requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
flickrapi==2.4.0
idna==3.10
oauthlib==3.2.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.3
click
